Move report routers into one folder
